requirements.txt
pyyaml>=6.0
pandas>=2.0
requests>=2.31
tqdm>=4.66
tabulate>=0.9

